Match cardread length to its 16-bit wire width
osdp_event_cardread::length carries a 16-bit value on the wire but was
declared int. Narrow it to uint16_t.

The bits-to-bytes conversion has to be widened at the same time: a
uint16_t operand promotes to int, and 0xFFFF + 7 overflows an int that
is only 16 bits wide on some targets. Fold it into a BITS_TO_BYTES()
macro that computes in uint32_t, and use it on both the encode and
decode paths.

+++ include/osdp.h
@@ -916,9 +916,9 @@ struct osdp_event_cardread {
 	 */
 	int direction;
 	/**
-	 * Length of card data in bits
+	 * Length of card data in bits. Carried as a 16-bit value on the wire.
 	 */
-	int length;
+	uint16_t length;
 	/**
 	 * Card data of @a length bytes or bits bits depending on @a format
 	 */
